Moved deploy plugin classes into main plugin module
