+++ plugins/core/camel/src/main/java/org/kantega/respiro/camel/CamelDeployPlugin.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.kantega.respiro.cameldeploy;
+package org.kantega.respiro.camel;
 
 import org.kantega.respiro.camel.CamelRouteDeployer;
 import org.apache.camel.builder.RouteBuilder;
