--- plugins/core/cxf-deploy/src/main/java/org/kantega/respiro/cxfdeploy/CxfDeployPlugin.java
@@ -1,41 +1,0 @@
-/*
- * Copyright 2015 Kantega AS
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.kantega.respiro.cxfdeploy;
-
-import org.kantega.respiro.api.EndpointConfig;
-import org.kantega.respiro.cxf.api.EndpointDeployer;
-import org.kantega.reststop.api.Plugin;
-
-import java.util.Collection;
-
-/**
- *
- */
-@Plugin
-public class CxfDeployPlugin  {
-
-
-    private final EndpointDeployer endpointDeployer;
-    private final Collection<EndpointConfig> endpointConfigs;
-
-    public CxfDeployPlugin(EndpointDeployer endpointDeployer, Collection<EndpointConfig> endpointConfigs) {
-        this.endpointDeployer = endpointDeployer;
-        this.endpointConfigs = endpointConfigs;
-        endpointDeployer.deployEndpoints(endpointConfigs);
-    }
-
-}
