+++ plugins/core/jersey/src/main/java/org/kantega/respiro/jersey/JerseyDeployPlugin.java
